+++ rollup.config.js
@@ -1,12 +1,13 @@
-import typescript from 'rollup-plugin-typescript2';
-import resolve from 'rollup-plugin-node-resolve';
+import typescript from '@rollup/plugin-typescript';
+import { nodeResolve } from '@rollup/plugin-node-resolve';
 import { terser } from 'rollup-plugin-terser';
-const pkg = require('./package.json');
+import pkg from './package.json'
+
 let version = pkg.version;
 const banner = `/* @OpenAyame/ayame-web-sdk@${version} */`;
 const plugins = [
- 	typescript(),
-  resolve({ browser: true })
+ 	typescript({ tsconfig: './tsconfig.json' }),
+  nodeResolve({ browser: true })
 ];
 module.exports = [
   {
@@ -14,11 +15,10 @@ module.exports = [
     output: {
       banner: banner,
       name: 'Ayame',
-      entry: 'src/',
       compact: true,
       file: 'dist/ayame.min.js',
       format: 'umd',
-      sourceMap: true
+      sourcemap: true
     },
     plugins: [...plugins, terser()]
   },
@@ -27,11 +27,20 @@ module.exports = [
     output: {
       banner: banner,
       name: 'Ayame',
-      entry: 'src/',
       file: 'dist/ayame.js',
       format: 'umd',
-      sourceMap: true
+      sourcemap: true
     },
     plugins: plugins
+  },
+  {
+    input: 'src/ayame.ts',
+    output: {
+      banner: banner,
+      name: 'Ayame',
+      file: 'dist/ayame.mjs',
+      format: 'module',
+    },
+    plugins: plugins,
   }
 ];